% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_functions.R
\name{add_lmsdata}
\alias{add_lmsdata}
\title{Add lmsdata}
\usage{
add_lmsdata(df)
}
\arguments{
\item{df}{Dataframe of new LMS data to add}
}
\description{
Function to add to lmsdata in data_env environment
}
