% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_functions.R
\name{get_chartmetadata}
\alias{get_chartmetadata}
\title{Get chart metadata}
\usage{
get_chartmetadata()
}
\description{
Function to return the dataframe containing chart metadata, including chart name, anthropometric measure and units,
age label and units with min, max, and range, number of parameters, source text and URL, and notes
}
\examples{
head(get_chartmetadata())
}
