% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_functions.R
\name{get_lms}
\alias{get_lms}
\title{Get LMS parameters function}
\usage{
get_lms(age, gender, chart = "fenton_2013", measure = "weight")
}
\arguments{
\item{age}{Vector of ages}

\item{gender}{Vector of gender, either 'm' or 'f'}

\item{chart}{Which chart to obtain LMS parameters for. Defaults to 'Fenton 2013' premature growth chart}

\item{measure}{Which measure for the chart. Defaults to 'weight'}
}
\description{
Function to interpolate LMS parameters, given age, gender, chart, and measure
}
\examples{
get_lms(c(38, 38), c('m', 'f'), chart = 'fenton_2013', measure = 'weight')
}
