% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_functions.R
\name{get_lmsdata}
\alias{get_lmsdata}
\title{Get lmsdata}
\usage{
get_lmsdata()
}
\description{
Function to return the dataframe containing all LMS values, measures, and units of measure, including any added with add_lmsdata()
}
\examples{
head(get_lmsdata())
}
