% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_functions.R
\name{lmsToValue}
\alias{lmsToValue}
\title{Z and LMS to X function}
\usage{
lmsToValue(L, M, S, Z)
}
\arguments{
\item{L}{lambda parameter}

\item{M}{mu parameter}

\item{S}{sigma parameter}

\item{Z}{Z parameter}
}
\description{
Vectorized function to convert Z + LMS parameters to a measurement
}
\examples{
# lmsToValue()
}
\keyword{LMS}
