% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_functions.R
\name{lmsToZ}
\alias{lmsToZ}
\title{X and LMS to Z function}
\usage{
lmsToZ(x, lms)
}
\arguments{
\item{x}{Vector of measurements}

\item{lms}{List of L, M, and S elements, each with length(x) elements}
}
\description{
function to convert X + LMS parameters to a Z-score
}
\examples{
# lmsToZ()
}
\keyword{LMS}
