% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_von.R
\name{recode_von_xml}
\alias{recode_von_xml}
\title{VON XML recoding function}
\usage{
recode_von_xml(von_xml)
}
\arguments{
\item{von_xml}{A link to the XML output of a VON data export}
}
\description{
This function takes a link to a VON XML export and coverts it into a dataframe and recodes the variables.
}
\examples{
# recode_von_xml()
}
\keyword{VON}
