% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_functions.R
\name{x_lms_to_z}
\alias{x_lms_to_z}
\title{X and LMS to Z function}
\usage{
x_lms_to_z(x, lms)
}
\arguments{
\item{x}{Vector of measurements}

\item{lms}{List of L, M, and S elements, each with length(x) elements}
}
\description{
function to convert X + LMS parameters to a Z-score
}
\examples{
# x_lms_to_z()
}
\keyword{LMS}
