% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_functions.R
\name{x_to_z}
\alias{x_to_z}
\title{X to Z-score function}
\usage{
x_to_z(x, age, gender, chart, measure = "weight")
}
\arguments{
\item{x}{Vector of measurements}

\item{age}{Vector of age parameters}

\item{gender}{Vector of genders, either 'm' or 'f'}

\item{chart}{Uniquely specified chart to obtain LMS parameters for}

\item{measure}{Uniquely specified measure for the chart. Defaults to 'weight'.}
}
\description{
Function to take vectors of measurements, age, and gender, and unique chart and measure, to return Z score
}
\details{
The following charts are available, with their corresponding measures.

\tabular{lll}{
  \strong{chart} \tab \strong{age_units} \tab \strong{measures} \cr
  abdel-rahman_2017     \tab months \tab arm_circ(cm)                                            \cr
  addo_2010_skin        \tab years  \tab subscapular(mm), triceps(mm)                            \cr
  addo_2017             \tab months \tab afa(cm2), ama(cm2), arm_circ(cm)                        \cr
  brooks_gmfcs_1        \tab years  \tab bmi(kg/m2), height(cm), weight(kg)                      \cr
  brooks_gmfcs_2        \tab years  \tab bmi(kg/m2), height(cm), weight(kg)                      \cr
  brooks_gmfcs_3        \tab years  \tab bmi(kg/m2), height(cm), weight(kg)                      \cr
  brooks_gmfcs_4        \tab years  \tab bmi(kg/m2), height(cm), weight(kg)                      \cr
  brooks_gmfcs_5_nt     \tab years  \tab bmi(kg/m2), height(cm), weight(kg)                      \cr
  brooks_gmfcs_5_tf     \tab years  \tab bmi(kg/m2), height(cm), weight(kg)                      \cr
  cappa_2024            \tab years  \tab height_velocity(cm/year), height(cm), weight(kg)        \cr
  cdc_2000_bmi          \tab months \tab bmi(kg/m2)                                              \cr
  cdc_2000_infant       \tab months \tab head_circ(cm), length(cm), weight(kg)                   \cr
  cdc_2000_pedi         \tab months \tab height(cm), weight(kg)                                  \cr
  cdc_2000_wt_for_ht    \tab cm     \tab weight(kg)                                              \cr
  cdc_2000_wt_for_len   \tab cm     \tab weight(kg)                                              \cr
  fenton_2003           \tab weeks  \tab head_circ(cm), length(cm), weight(kg)                   \cr
  mramba_2017           \tab months \tab arm_circ(cm)                                            \cr
  olsen_2010            \tab weeks  \tab head_circ(cm), length(cm), weight(g)                    \cr
  olsen_2015_bmi        \tab weeks  \tab bmi(g/cm2)                                              \cr
  who_2006_infant       \tab months \tab head_circ(cm), length(cm), weight(kg)                   \cr
  who_2007_skin_arm     \tab months \tab arm_circ(cm), subscapular(mm), triceps(mm)              \cr
  who_expanded          \tab days   \tab bmi(kg/m2), head_circ(cm), length_height(cm), weight(kg)\cr
  who_expanded_arm_skin \tab days   \tab arm_circ(cm), subscapular(mm), triceps(mm)              \cr
  who_wt_for_ht         \tab cm     \tab weight(kg)                                              \cr
  who_wt_for_len        \tab cm     \tab weight(kg)                                              \cr
  zemel_2015_infant     \tab months \tab head_circ(cm), length(cm), weight(kg)                   \cr
  zemel_2015_pedi       \tab years  \tab bmi(kg/m2), head_circ(cm), height(cm), weight(kg)       
}
}
\examples{
# 3, 10, 50, 90, and 97\%ile weight for 8.5 month old female on CDC 2000 infant chart
x_to_z(
    c(6.720327734, 7.197413532, 8.314178377, 9.573546299, 10.2153883),
    rep(8.5, 5),
    rep('f', 5),
    chart = 'cdc_2000_infant',
    measure = 'weight'
  )
round( pnorm( x_to_z(
    c(6.720327734, 7.197413532, 8.314178377, 9.573546299, 10.2153883),
    rep(8.5, 5),
    rep('f', 5),
    chart = 'cdc_2000_infant',
    measure = 'weight'
  ) ), 4)
}
