% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_functions.R
\name{xToZ}
\alias{xToZ}
\title{X to Z-score function}
\usage{
xToZ(x, age, gender, chart = "Fenton2013", measure = "Wt")
}
\arguments{
\item{x}{Vector of measurements}

\item{age}{Vector of age parameters}

\item{gender}{Vector of genders, either 'M' or 'F'}

\item{chart}{Uniquely specified chart to obtain LMS parameters for. Defaults to 'Fenton 2013' premature growth chart.}

\item{measure}{Uniquely specified measure for the chart. Defaults to 'Wt'.}
}
\description{
Function to take a vectors of measurements, age, and gender, and unique chart and measure, to return Z score
}
\examples{
# 3, 10, 50, 90, and 97\%ile for 30 0/7 week M on Fenton2013
xToZ(
    c(774.1259148, 986.5784793, 1388.303356, 1746.218584, 1903.278428),
    rep(30, 5),
    rep('M', 5),
    chart = 'Fenton2013',
    measure = 'Wt'
  )
round( pnorm( xToZ(
    c(774.1259148, 986.5784793, 1388.303356, 1746.218584, 1903.278428),
    rep(30, 5),
    rep('M', 5),
    chart = 'Fenton2013',
    measure = 'Wt'
  ) ), 4)
}
\keyword{LMS}
