% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_functions.R
\name{z_lms_to_x}
\alias{z_lms_to_x}
\title{Z and LMS to X function}
\usage{
z_lms_to_x(Z, L, M, S)
}
\arguments{
\item{Z}{Z parameter(s)}

\item{L}{lambda parameter(s)}

\item{M}{mu parameter(s)}

\item{S}{sigma parameter(s)}
}
\description{
Vectorized function to convert Z + LMS parameters to a measurement. If Z, L, M, or S are vectors of different length, will recycle the shorter vectors.
}
\examples{
# z_lms_to_x()
}
