% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_functions.R
\name{get_lms}
\alias{get_lms}
\title{Get LMS parameters function}
\usage{
get_lms(age, gender, chart, measure = "weight")
}
\arguments{
\item{age}{Vector of ages}

\item{gender}{Vector of gender, either 'm' or 'f'}

\item{chart}{Which chart to obtain LMS parameters for}

\item{measure}{Which measure for the chart. Defaults to 'weight'}
}
\description{
Function to interpolate LMS parameters, given age, gender, chart, and measure
}
\examples{
get_lms(c(11.5, 11.5), c('m', 'f'), chart = 'cdc_2000_infant', measure = 'weight')
}
