% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_functions.R
\name{get_lms}
\alias{get_lms}
\title{Get LMS parameters function}
\usage{
get_lms(age, gender, chart = "Fenton2013", measure = "Wt")
}
\arguments{
\item{age}{Vector of ages}

\item{gender}{Vector of gender, either 'M' or 'F'}

\item{chart}{Which chart to obtain LMS parameters for. Defaults to 'Fenton 2013' premature growth chart}

\item{measure}{Which measure for the chart. Defaults to 'Wt'}
}
\description{
Function to interpolate LMS parameters, given age, gender, chart, and measure
}
\examples{
get_lms(c(38, 38), c('M', 'F'), chart = 'Fenton2013', measure = 'Wt')
}
