% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_von.R
\name{recode_von}
\alias{recode_von}
\title{VON recoding function}
\usage{
recode_von(von)
}
\arguments{
\item{von}{A dataframe output of a VON data export}
}
\description{
This function takes a VON export converted into a dataframe, and recodes the variables.
}
\examples{
# recode_von()
}
\keyword{VON}
