% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_functions.R
\name{x_to_z}
\alias{x_to_z}
\title{X to Z-score function}
\usage{
x_to_z(x, age, gender, chart = "Fenton2013", measure = "Wt")
}
\arguments{
\item{x}{Vector of measurements}

\item{age}{Vector of age parameters}

\item{gender}{Vector of genders, either 'M' or 'F'}

\item{chart}{Uniquely specified chart to obtain LMS parameters for. Defaults to 'Fenton 2013' premature growth chart.}

\item{measure}{Uniquely specified measure for the chart. Defaults to 'Wt'.}
}
\description{
Function to take vectors of measurements, age, and gender, and unique chart and measure, to return Z score
}
\details{
The following charts are available, with their corresponding measures.

\tabular{lll}{
  \strong{Chart} \tab \strong{Age} \tab \strong{Measures}\cr
  CDCbmi \tab months \tab BMI(kg/m2)\cr
  CDCinfant \tab months \tab HC(cm), Len(cm), Wt(kg)\cr
  CDCpedi \tab months \tab Len(cm), Wt(kg)\cr
  CDCskin \tab years \tab subscapular(mm), triceps(mm)\cr
  DownInfant \tab months \tab HC(cm), Len(cm), Wt(kg)\cr
  DownPedi \tab years \tab BMI(kg/m2), HC(cm), Height(cm), Wt(kg)\cr
  Fenton2003 \tab weeks \tab HC(cm), Len(cm), Wt(kg)\cr
  Olsen2010 \tab weeks \tab HC(cm), Len(cm), Wt(g)\cr
  OlsenBMI \tab weeks \tab BMI(g/cm2)\cr
  WHOinfant \tab months \tab HC(cm), Len(cm), Wt(kg)\cr
  WHOskin \tab months \tab arm circ(cm), subscapular(mm), triceps(mm)\cr
}
}
\examples{
# 3, 10, 50, 90, and 97\%ile for 30 0/7 week M on Fenton2013
x_to_z(
    c(774.1259148, 986.5784793, 1388.303356, 1746.218584, 1903.278428),
    rep(30, 5),
    rep('M', 5),
    chart = 'Fenton2013',
    measure = 'Wt'
  )
round( pnorm( x_to_z(
    c(774.1259148, 986.5784793, 1388.303356, 1746.218584, 1903.278428),
    rep(30, 5),
    rep('M', 5),
    chart = 'Fenton2013',
    measure = 'Wt'
  ) ), 4)
}
